Crate fpma

fpma ia a small cli/library to parse and plot nhmmer.table files generated by the flag --tblout on nhmmer.

See the plot module for plotting details.

See the cli module for the command line interface details.

The gff module creates GFF3 files from the nhmmer.table files.

For an example of what the plot looks like, please see this example.

Modules

cli

A module to parse command line args.

gff

A lightweight GFF3 writer.

plot

A module to plot the output of nhmmer (HMMER3).

Structs

Nhmmer

Database representation of the tabular output of nhmmer.

NhmmerRow

A row representation of the tabular output of nhmmer using the --tblout option.

Enums

MitoGene

An enumeration of all the genes checked for in this pipeline.

Strand

The strandedness of the HMM hit in the genome.

Functions

run_hmmer

Entry point for running hmmer.