Enum fpma::MitoGene

pub enum MitoGene {

    Atp1,
    Atp4,
    Atp6,
    Atp8,
    Atp9,
    CcmB,
    CcmC,
    CcmFc,
    CcmFn,
    Cob,
    Cox1,
    Cox2,
    Cox3,
    MatR,
    MttB,
    Nad1,
    Nad2,
    Nad3,
    Nad4,
    Nad4L,
    Nad5,
    Nad6,
    Nad7,
    Nad9,
    Rpl2,
    Rpl5,
    Rpl6,
    Rpl10,
    Rpl16,
    Rps1,
    Rps2,
    Rps3,
    Rps4,
    Rps7,
    Rps8,
    Rps10,
    Rps11,
    Rps12,
    Rps13,
    Rps14,
    Rps19,
    Sdh3,
    Sdh4,
    TrnA_UGC,
    TrnC_GCA,
    TrnD_GUC,
    TrnE_UUC,
    TrnF_GAA,
    TrnG_GCC,
    TrnG_UCC,
    TrnH_GUG,
    TrnI_CAU,
    TrnI_GAU,
    TrnK_UUU,
    TrnL_CAA,
    TrnL_GAG,
    TrnL_UAA,
    TrnL_UAG,
    TrnfM_CAU,
    TrnM_CAU,
    TrnN_GUU,
    TrnP_UGG,
    TrnQ_UUG,
    TrnR_ACG,
    TrnR_UCG,
    TrnR_UCU,
    TrnS_GCU,
    TrnS_GGA,
    TrnS_UGA,
    TrnT_UGU,
    TrnV_GAC,
    TrnV_UAC,
    TrnW_CCA,
    TrnY_GUA,

}

An enumeration of all the genes checked for in this pipeline.

There are 43 core genes currently, some of which are not present in all of the plant clades.

In addition, there are 31 currently enumerated tRNA genes.

Variants

Atp1

Atp4

Atp6

Atp8

Atp9

CcmB

Lost in hornwort, lycopytes

CcmC

Lost in hornwort, lycopytes

CcmFc

Lost in hornwort, lycopytes

CcmFn

Lost in hornwort, lycopytes

Cob

Cox1

Cox2

Cox3

MatR

MttB

Nad1

Nad2

Nad3

Nad4

Nad4L

Nad5

Nad6

Nad7

Nad9

Rpl2

Rpl5

Rpl6

Not yet recorded in angiosperms.

Rpl10

Rpl16

Rps1

Rps2

Rps3

Rps4

Rps7

Rps8

Not yet recorded in angiosperms.

Rps10

Rps11

Rps12

Rps13

Rps14

Rps19

Sdh3

Sdh4

TrnA_UGC

Potentially absent from Seed plants

TrnC_GCA

TrnD_GUC

TrnE_UUC

TrnF_GAA

TrnG_GCC

TrnG_UCC

Potentially absent from Seed plants

TrnH_GUG

Potentially absent from Seed plants

TrnI_CAU

TrnI_GAU

Potentially absent from Seed plants

TrnK_UUU

TrnL_CAA

Potentially absent from Seed plants

TrnL_GAG

Potentially absent from Seed plants

TrnL_UAA

Potentially absent from Seed plants

TrnL_UAG

Potentially absent from Seed plants

TrnfM_CAU

TrnM_CAU

Potentially absent from Seed plants

TrnN_GUU

Potentially absent from Seed plants

TrnP_UGG

TrnQ_UUG

TrnR_ACG

Potentially absent from Seed plants

TrnR_UCG

Potentially absent from Seed plants

TrnR_UCU

Potentially absent from Seed plants

TrnS_GCU

TrnS_GGA

Potentially absent from Seed plants

TrnS_UGA

TrnT_UGU

Potentially absent from Seed plants

TrnV_GAC

Potentially absent from Seed plants

TrnV_UAC

TrnW_CCA

TrnY_GUA

Implementations

impl MitoGene

pub fn iterator() -> Iter<'static, MitoGene>

Returns an iterator over the MitoGenes.