Module fpma::plot

A module to plot the output of nhmmer (HMMER3).

Structs

PlotData

PlotData database. Composed of a BTreeMap, where the keys are the contigs/fasta ID’s and the values are a vector of PlotDataRow.

PlotDataRow

PlotData row entry.

Statics

MARGIN

Size of the margins in the plot.

SUBPLOT_HEIGHT

The height of each subplot.

WIDTH

The width of the plot.

Functions

format_axis_label_len

Add bp to the end of a string.

format_bp_pretty

Pretty print i32 numbers.

generate_plot_annotations

A function which returns the arrows representing genes along the axis of the mitochondrial contig.

make_html

Function to make the HTML string.

scale_x

Scale an x value from the data scale to the visualisation scale.